services/gpu-embedding/requirements.txt
# GPU Embedding Service - Requirements
# Python 3.10+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GPU & Deep Learning
torch==2.1.0
torchvision==0.16.0
sentence-transformers==2.2.2
transformers==4.35.0

# Vector Search
faiss-gpu==1.7.4

# Utilities
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
httpx==0.25.1

# Monitoring
prometheus-client==0.18.0
python-json-logger==2.0.7

# Environment
python-dotenv==1.0.0
